Split cloud credential recovery tests
